migrate actant table to render
