+++ packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditorActantTable/StatementEditorActantTable.tsx
@@ -1,16 +1,12 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { AttributeIcon } from "components";
-import { Column, useTable, useExpanded, Row, Cell } from "react-table";
+import React, { useCallback, useMemo, useState } from "react";
+import { Column, useTable, useExpanded, Row } from "react-table";
 import update from "immutability-helper";
 import {
   StyledTable,
   StyledTHead,
   StyledTh,
-  StyledTagWrapper,
 } from "./StatementEditorActantTableStyles";
 import { StatementEditorActantTableRow } from "./StatementEditorActantTableRow";
-// import { AttributesEditor } from "../../AttributesEditor/AttributesEditor";
-import { AttributeButtonGroup } from "../../AttributeButtonGroup/AttributeButtonGroup";
 
 import {
   IActant,
@@ -18,13 +14,8 @@ import {
   IStatementActant,
   IStatementProp,
 } from "@shared/types";
-import { EntitySuggester, EntityTag } from "../..";
-import { Button, ButtonGroup } from "components";
-import { FaTrashAlt, FaUnlink, FaPlus } from "react-icons/fa";
 import { UseMutationResult } from "react-query";
-import { excludedSuggesterEntities } from "Theme/constants";
 import { ActantType } from "@shared/enums";
-import AttributesEditor from "../../AttributesEditor/AttributesEditor";
 
 interface FilteredActantObject {
   data: { actant: IActant | undefined; sActant: IStatementActant };
@@ -72,25 +63,6 @@ export const StatementEditorActantTable: React.FC<
     setFilteredActants(filteredActants);
   }, [statement]);
 
-  const updateActant = (statementActantId: string, changes: any) => {
-    if (statement && statementActantId) {
-      const updatedActants = statement.data.actants.map((a) =>
-        a.id === statementActantId ? { ...a, ...changes } : a
-      );
-      const newData = { actants: updatedActants };
-      updateActantsMutation.mutate(newData);
-    }
-  };
-  const removeActant = (statementActantId: string) => {
-    if (statement) {
-      const updatedActants = statement.data.actants.filter(
-        (a) => a.id !== statementActantId
-      );
-      const newData = { actants: updatedActants };
-      updateActantsMutation.mutate(newData);
-    }
-  };
-
   const updateActantsOrder = () => {
     if (userCanEdit) {
       const actants: IStatementActant[] = filteredActants.map(
@@ -111,188 +83,13 @@ export const StatementEditorActantTable: React.FC<
       {
         Header: "Actant",
         accessor: "data",
-        Cell: ({ row }: Cell) => {
-          const {
-            actant,
-            sActant,
-          }: {
-            actant: IActant;
-            sActant: IStatementActant | any;
-          } = row.values.data;
-          return actant ? (
-            <StyledTagWrapper>
-              <EntityTag
-                actant={actant}
-                // fullWidth
-                button={
-                  userCanEdit && (
-                    <Button
-                      key="d"
-                      tooltip="unlink actant"
-                      icon={<FaUnlink />}
-                      color="plain"
-                      inverted={true}
-                      onClick={() => {
-                        updateActant(sActant.id, {
-                          actant: "",
-                        });
-                      }}
-                    />
-                  )
-                }
-              />
-            </StyledTagWrapper>
-          ) : (
-            userCanEdit && (
-              <EntitySuggester
-                onSelected={(newSelectedId: string) => {
-                  updateActant(sActant.id, {
-                    actant: newSelectedId,
-                  });
-                }}
-                categoryTypes={classEntitiesActant}
-                excludedEntities={excludedSuggesterEntities}
-              />
-            )
-          );
-        },
       },
       {
         id: "position",
         Header: "Position",
-        Cell: ({ row }: Cell) => {
-          const { sActant } = row.values.data;
-          return (
-            <AttributeButtonGroup
-              disabled={!userCanEdit}
-              options={[
-                {
-                  longValue: "subject",
-                  shortValue: "s",
-                  onClick: () =>
-                    updateActant(sActant.id, {
-                      position: "s",
-                    }),
-                  selected: sActant.position == "s",
-                },
-                {
-                  longValue: "actant1",
-                  shortValue: "a1",
-                  onClick: () =>
-                    updateActant(sActant.id, {
-                      position: "a1",
-                    }),
-                  selected: sActant.position == "a1",
-                },
-                {
-                  longValue: "actant2",
-                  shortValue: "a2",
-                  onClick: () =>
-                    updateActant(sActant.id, {
-                      position: "a2",
-                    }),
-                  selected: sActant.position == "a2",
-                },
-                {
-                  longValue: "pseudo-actant",
-                  shortValue: "pa",
-                  onClick: () =>
-                    updateActant(sActant.id, {
-                      position: "p",
-                    }),
-                  selected: sActant.position == "p",
-                },
-              ]}
-            />
-          );
-        },
       },
       {
         id: "Attributes",
-        Cell: ({ row }: Cell) => {
-          const {
-            actant,
-            sActant,
-          }: {
-            actant: IActant;
-            sActant: IStatementActant | any;
-          } = row.values.data;
-          const propOriginId = row.values.data.sActant.actant;
-          return (
-            <ButtonGroup noMargin>
-              {sActant && (
-                <AttributesEditor
-                  modalTitle={`Actant involvement`}
-                  actant={actant}
-                  disabledAllAttributes={!userCanEdit}
-                  userCanEdit={userCanEdit}
-                  data={{
-                    elvl: sActant.elvl,
-                    certainty: sActant.certainty,
-                    logic: sActant.logic,
-                    virtuality: sActant.virtuality,
-                    partitivity: sActant.partitivity,
-                    operator: sActant.operator,
-                    bundleStart: sActant.bundleStart,
-                    bundleEnd: sActant.bundleEnd,
-                  }}
-                  handleUpdate={(newData) => {
-                    updateActant(sActant.id, newData);
-                  }}
-                  updateActantId={(newId: string) => {
-                    updateActant(sActant.id, { actant: newId });
-                  }}
-                  classEntitiesActant={classEntitiesActant}
-                  loading={updateActantsMutation.isLoading}
-                />
-              )}
-              {userCanEdit && (
-                <Button
-                  key="d"
-                  icon={<FaTrashAlt />}
-                  color="plain"
-                  inverted={true}
-                  tooltip="remove actant row"
-                  onClick={() => {
-                    removeActant(row.values.data.sActant.id);
-                  }}
-                />
-              )}
-              {userCanEdit && (
-                <Button
-                  key="a"
-                  icon={<FaPlus />}
-                  color="plain"
-                  inverted={true}
-                  tooltip="add new prop"
-                  onClick={() => {
-                    addProp(propOriginId);
-                  }}
-                />
-              )}
-              {sActant.logic == "2" && (
-                <Button
-                  key="neg"
-                  tooltip="Negative logic"
-                  color="success"
-                  inverted={true}
-                  noBorder
-                  icon={<AttributeIcon attributeName={"negation"} />}
-                />
-              )}
-              {sActant.operator && (
-                <Button
-                  key="oper"
-                  tooltip="Logical operator type"
-                  color="success"
-                  inverted={true}
-                  noBorder
-                  icon={sActant.operator}
-                />
-              )}
-            </ButtonGroup>
-          );
-        },
       },
     ];
   }, [filteredActants, updateActantsMutation.isLoading]);
@@ -364,6 +161,9 @@ export const StatementEditorActantTable: React.FC<
                   userCanEdit={userCanEdit}
                   updateOrderFn={updateActantsOrder}
                   visibleColumns={visibleColumns}
+                  classEntitiesActant={classEntitiesActant}
+                  updateActantsMutation={updateActantsMutation}
+                  addProp={addProp}
                   {...row.getRowProps()}
                 />
               );
